test:전략 패턴을 위해 덧셈 테스트를 StringCalculator에서 AddOperator로 리팩토링
